Split `outbound::tcp::logical::tests` into a file
The TCP logical stack tests are nearly twice a long as the TCP logical
stack. This change moves the tests module into a distinct file for
readability.
